requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dj-config-url==0.1.1
Django==5.1.6
django-debug-toolbar==5.2.0
djangorestframework==3.15.2
dotenv==0.9.9
drf-spectacular==0.28.0
factory_boy==3.3.3
Faker==37.5.3
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==24.2
pillow==11.1.0
pluggy==1.6.0
psycopg2==2.9.10
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.27.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
